+++ src/main/java/org/parabot/core/parsers/servers/ServerParser.java
@@ -2,7 +2,7 @@ package org.parabot.core.parsers.servers;
 
 import org.parabot.core.Core;
 import org.parabot.core.desc.ServerDescription;
-import org.parabot.environment.servers.executers.ServerExecuter;
+import org.parabot.environment.servers.executers.ServerExecutor;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -15,7 +15,7 @@ import java.util.TreeMap;
  * @author Everel
  */
 public abstract class ServerParser {
-    public static final Map<ServerDescription, ServerExecuter> SERVER_CACHE = new HashMap<>();
+    public static final Map<ServerDescription, ServerExecutor> SERVER_CACHE = new HashMap<>();
 
     public static final ServerDescription[] getDescriptions() {
         SERVER_CACHE.clear();
@@ -41,7 +41,7 @@ public abstract class ServerParser {
             Core.verbose("Server providers parsed.");
         }
 
-        Map<ServerDescription, ServerExecuter> SORTED_SERVER_CACHE = new TreeMap<ServerDescription, ServerExecuter>(SERVER_CACHE);
+        Map<ServerDescription, ServerExecutor> SORTED_SERVER_CACHE = new TreeMap<ServerDescription, ServerExecutor>(SERVER_CACHE);
 
         return SORTED_SERVER_CACHE.keySet().toArray(new ServerDescription[SORTED_SERVER_CACHE.size()]);
     }
